split align command in 3 classes/commands
 for cookbooks, data bags and roles
